Configure Vercel deployment with serverless API and reminder cron
- Extract Express app construction into createApp() (server/_core/app.ts)
- Add api/index.ts as Vercel serverless handler wrapping the Express app
- Add vercel.json with /api/* rewrites and a 15-min reminder cron
- Add /api/cron/reminders endpoint guarded by CRON_SECRET so the
  scheduled reminder loop survives the move from setInterval
- Document required env vars and known limitations (file uploads,
  hourly auto-promotion) in the README

The dev workflow (pnpm dev) is unchanged: server/_core/index.ts now
imports createApp() but still owns the local listen() and setInterval
background jobs.
