+++ .gitignore
@@ -1,0 +1,9 @@
+node_modules/
+.pnpm-store/
+dist/
+build/
+.env
+.env.local
+.DS_Store
+*.log
+.vercel
